🏇 Refrain from zeroing the input buffer until needed
Previously, the input buffer was being zeroed upon initialization
and after being XORed into the state. Because bytes in the input
buffer are overwritten, this results in a performance hit.

Instead, the remaining bytes in the input buffer are zeroed when
digesting prior to padding.

The downside, however, is that some portion of the previous input
block remains in memory even after the next block begins.