🏇 Attempt to improve performance by precomputing permutation logic
The goal of this commit is two-fold:

 1. Unroll loops.
 2. Minimize array lookups.

To achieve this, the `permute()` function and each of its stages was
rewritten as a *code generator* rather than a direct implementation.
All deterministic computations and loops are replaced by direct
statements.

The result is a 10,000+ line output file containing the precomputed
`permute()` function.

The results are disappointingly on par with the direct implementation.
This is only a first pass, but it doesn't look like there will be an
easy win here.